% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade-discs.r
\name{cascade_disks}
\alias{cascade_disks}
\title{cascade_disks}
\usage{
cascade_disks(center, spacings, quadsegs)
}
\arguments{
\item{center}{SpatialPoints or SpatialPointsDataFrame object as defined in package sp}

\item{spacings}{vector giving the width of each ring}

\item{quadsegs}{Number of line segments to use to approximate a quarter circle.}
}
\value{
a SpatialPointsDataFrame with a unqiue attribute \code{id}
}
\description{
Creates SpatialPolygonDataFrame for a suite of imbriqued disks
}
\examples{
library(sp)
pt <- SpatialPoints(data.frame(x = 1000000 * runif(1), y = 1000000 * runif(1)))
res <- cascade_disks(center = pt, spacings = 10)
plot(res, col = "lightgrey") 
plot(pt, add = TRUE, col = 2)

}
\author{
Pierre Roudier
}

