% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-disc.r
\name{split_disc}
\alias{split_disc}
\title{split_disc}
\usage{
split_disc(center, width, splits)
}
\arguments{
\item{center}{a object of class \code{SpatialPoints} or \code{SpatialPointsDataFrame}}

\item{width}{radius of the disc to split into wedges}

\item{splits}{number of splits/wedges}
}
\value{
An object of class \code{SpatialPolygonsDataFrame}
}
\description{
Splits a disc into wedges
}
\examples{
pt <- SpatialPoints(data.frame(x = 1000000 * runif(1), y = 1000000 * runif(1)))
splits <- split_disc(center = pt, radius = 100, splits = 50)

plot(splits, col = "lightgrey")
plot(pt, add = TRUE, pch = "+", col = 2)

}
\author{
Pierre Roudier
}

